Fold all resource partials under a sub-directory in `/sources/includes` (#360)
With the documentation expanding to accommodate Booqable apps it's good
to be a little more tidy about the folder structure of this repo. This
PR moves the resource docs under their own folder.

There will be a small time period where updates in the resources won't
show up in the built website because the generator script in the main
Booqable repo won't yet be updated to dump the files in the correct
subfolder.

Counterpart to https://github.com/booqable/booqable/pull/14195.